requirements.txt
customtkinter==5.2.0
Pillow==10.0.1
tkinterdnd2==0.3.0
opencv-python==4.8.1.78
imageio==2.31.5
pyinstaller==6.1.0
packaging